👔 judge→janken
